Big changes made to project structure regarding backend.
Created new package containing four classes. These classes all utilise
DatabaseInteract class that interacts with the databse. The 4 classes
each have the purpose of implementing methods to be used in the 4 views
(Customer, Kitchen, Waiter, and Login). For example, the customer view
will use the CustomerAccess Methods such as getMenu(), and placeOrder().
This reduces the complexity of the project greatly and will help
massively when intergrating the entire project later.