Connected Log in system with databse.
Added a text file authentication that will contain staff information
(id, password, role). This is now inputed into the database. Methods in
loginAccess class now have correct functionality to check if user exists
and has a correct password, and then checks the role of the staff. Also
had to alter one method in database class to implement this.