Order constructor now uses String as the data type for items.
This has been done so that consumables can be represented more easily.
Currently, the constructor takes the string, splits it by using commas
as a delimiter and stores each element in an array. A price between
£0-15 is randomly generated for it and the string element and price is
assigned to a new Consumable object which is added to the Order's items
list field.