Implemented method that once a button is pushed updates Alerts.
Waiter view- added button and listpane into Alert tab. When the button
is pushed this will call a new methos -reloadAlert- that will access the
databse and retrieve the currently stored messages from customers. It
will then display this in the listpane. This is the bare bones method
and later will require updates but the basic functionality is working.