Added methods to Backend package in KitchenAccess class.
These methods send and recieve messages from the database so that the
mitchen can send messages to the waiter and recieve messages back to be
represented in the gui. The actual SQL queries are currently commented
out due to the database being down, but once the database is back up and
running the queries will be uncommeneted and should be fully functional.