The kitchenview now displays messages from waiters.
KitchenView notify tab, on clicking the button get messsages it will now
display messages from the waiter in a listview. Currenctly the actual
connection to the databse is commented out but once the databse is back
up and running this code can be uncommented and will work with the
databse to display messages.