Addded two methods in WaiterAccess class to deal with payements.
Method getPayements gets the payements currently stored in the databse
and returns them as a resultset. The functionality is currently
commented out as the databse is down whilst it is being communicated
with the server. The second method confirmPayement removes a payement
from the database once it has been confirmed.