Waiter can now confirm a table has paid in Waiter View.
The Waiterview now aollows selecting a payement and confirming as paid.
I added the functionality to complete this in WaiterViewCOntroller in
the confirm method, where the selected value in listview os removed and
the mehtod that removes the payement from the databse is called (for now
the database method is commented out due to the databse being down).