Implemented a system to manually test parts of the project.
Many parts of the project cannot be tested with automatic tests. Many
apsects are visual or to do with complex features that would be very
hard to test. So to test more complicated parts of the project I have
created a manual test file (textfile) that will use a standard method
for testing major parts of the project manually and will be stored in
Manual_Testing textfile.