Restructured UserThread's initialisation method.
Both customer and staff's requestthreads (and notifcationthreads) will
now specify that they are such threads in their first message to the
server. UserThread's run method has now been split into three methods.
The original checks if thread is notification or request and calls the
respective method for each to check if the message comes from a customer
or staff.