Customer view controller establishes server connection on init.
The MainViewController now gets an instance of ServerAccess as a field.
When the view is initialised, it calls the setConnection method to
connect with the server, using the table number inserted on startup as
an identifier for the user. Because the controller has an instance of
ServerAccess, calling this method now also gives the controller instance
its own RequestThread and NotificationThread.