Added delegate getMenu method to ServerAccess.
MainViewController can now get the menu for a customer on initialisation
now a delegate method to call getMenu in RequestThread has been added to
ServerAccess. Also made the getInstance method in ServerAccess static in
accordance with the Singleton design pattern and fixed some programming
errors after changing the ServerAccess library imported from the staff's
version to the customers.