Added role verification in staff's RequestThread.
Since all actions that waiters and kitchen staff can do are stored in
one class, a field has been added  to RequestThread to store the role of
the user, based on what the server writes when a staff user logs in.
This field is then used to check that the user belongs to the role that
can do an action. For example, confirming an order now requires a check
to see if the user's role is waiter before proceeding.