Added conditions for cancelling order in staff RequestThread
The method now returns false if the order is complete. If this condition
isn't met, we check that, if the cancelling user has the kitchen role,
the order status isn't confirmed or processing, which also results in
returning false if this condition is met.

Also tidied up exception handling in each method of staff request thread
by having them return the result of checkAccepted and printing the stack
trace in the catch block.

Finally, the checkAccepted method has been added to the customer's
RequestThread as well as the same changes to exception handling
described above.