Rewritten staff's ServerAccess to resemble customer's ServerAccess.
With the updated methods of RequestThread and NotificationThread, there
is no need for ReadThread or WriteThread, they have been deleted. The
staff's ServerAccess class has been rewritten to work similarly to the
customer's, with all 10 functional methods from RequestThread and
NotificationThread included as delegation methods. Also, the type
parameter of staff's request/notification constructors has been removed
as it is not needed. An ID field has been added to both classes,
corresponding to the user's username, along with a getter method, used
when creating a new NotificationThread in ServerAccess's setConnection
method. getMenu method has also been added to staff's RequestThread.